requirements.txt
flask-restful
automl
flask~=1.1.2
google-cloud
six~=1.14.0
pip~=20.3.3
keyring~=18.0.1
rsa~=4.7
numpy~=1.17.4
nose~=1.3.7
pytz~=2019.3
PyYAML~=5.3.1
Jinja2~=2.11.2
Werkzeug~=1.0.1
click~=7.1.2
blinker~=1.4
setuptools~=51.3.3
cryptography~=2.8
urllib3~=1.25.8
requests~=2.22.0
MarkupSafe~=1.1.1
chardet~=3.0.4
Pillow~=7.0.0
simplejson~=3.16.0
idna~=2.8
api~=0.0.7
wheel~=0.36.2